requirements.txt
Flask==3.0.0
phonenumbers==8.13.25
pyyaml==6.0.1
